Producer API dwarf_add_frame_cie use Dwarf_Small(uint8_t) to hold daf (data alignment factor). Internally, daf is stored in a Dwarf_Signed(int64_t) data field of the CIE structure. As a result, we need to cast daf to int8_t before size promotion to keep signedness.
Note that according to DWARF spec (version 2 or 3), daf should be a
signed 128bit(encoded) value. This 8bit restriction was actually added
by SGI libdwarf, I think.

+++ libdwarf/dwarf_pro_frame.c
@@ -75,7 +75,7 @@ dwarf_add_frame_cie(Dwarf_P_Debug dbg, char *augmenter, Dwarf_Small caf,
 	}
 
 	cie->cie_caf = caf;
-	cie->cie_daf = daf;
+	cie->cie_daf = (int8_t) daf; /* daf is signed. */
 	cie->cie_ra = ra;
 	if (initinst != NULL && inst_len > 0) {
 		cie->cie_initinst = malloc(inst_len);
